requirements.txt
flask
flask_sqlalchemy
xlwt
xlrd
pdf2image
pdfschedule
beautifulsoup4
selenium
Flask-Bcrypt
Flask-Login
Flask-Mail
requests
flask_wtf
wtforms
itsdangerous==2.0.1
authlib
PyMuPDF
cachecontrol
email_validator
googlemaps
python-dotenv